♻️ refactor(pages): rename NewCustomScreen to ComposableScreen
- rename directories, files, schemas, types, and renderer symbols
- update literal discriminator from "NewCustomScreen" to "ComposableScreen"
- update all import paths and re-exports across both packages
