+++ packages/onboarding-ui/src/UI/OnboardingPage.tsx
@@ -1,5 +1,5 @@
 import { OnboardingStepType } from "./types";
-import { RatingsRenderer, PickerRenderer, CommitmentRenderer, CarouselRenderer, LoaderRenderer, MediaContentRenderer, NewCustomScreenRenderer, QuestionRenderer, QuestionAnswerButtonProps, QuestionAnswersListProps } from "./Pages";
+import { RatingsRenderer, PickerRenderer, CommitmentRenderer, CarouselRenderer, LoaderRenderer, MediaContentRenderer, ComposableScreenRenderer, QuestionRenderer, QuestionAnswerButtonProps, QuestionAnswersListProps } from "./Pages";
 import { View, Text, Button } from 'react-native';
 import { useTheme } from "./Theme/useTheme";
 import { Theme } from "./Theme";
@@ -34,8 +34,8 @@ export const OnboardingPage = ({ step, onContinue, isSandbox }: OnboardingPagePr
       return <LoaderRenderer step={step} onContinue={onContinue} theme={theme} />;
     case 'Question':
       return <QuestionRenderer step={step} onContinue={onContinue} theme={theme} />;
-    case 'NewCustomScreen':
-      return <NewCustomScreenRenderer step={step} onContinue={onContinue} theme={theme} />;
+    case 'ComposableScreen':
+      return <ComposableScreenRenderer step={step} onContinue={onContinue} theme={theme} />;
     default:
       if (isSandbox) {
         // @ts-ignore
