+++ packages/onboarding-ui/src/UI/Pages/index.ts
@@ -2,7 +2,7 @@ export * from "./Carousel";
 export * from "./Commitment";
 export * from "./Loader";
 export * from "./MediaContent";
-export * from "./NewCustomScreen";
+export * from "./ComposableScreen";
 export * from "./Picker";
 export * from "./Question";
 export * from "./Ratings";
