+++ packages/onboarding-ui/src/UI/types.ts
@@ -3,7 +3,7 @@ import {
   CommitmentStepType,
   LoaderStepType,
   MediaContentStepType,
-  NewCustomScreenStepType,
+  ComposableScreenStepType,
   PickerStepType,
   RatingsStepType,
   QuestionStepType,
@@ -34,7 +34,7 @@ export type BaseStepType = {
 export type OnboardingStepType =
   | RatingsStepType
   | MediaContentStepType
-  | NewCustomScreenStepType
+  | ComposableScreenStepType
   | PickerStepType
   | CommitmentStepType
   | CarouselStepType
