+++ packages/onboarding/src/steps/types.ts
@@ -2,7 +2,7 @@ import { CarouselStepType } from "./Carousel/types";
 import { CommitmentStepType } from "./Commitment/types";
 import { LoaderStepType } from "./Loader/types";
 import { MediaContentStepType } from "./MediaContent/types";
-import { NewCustomScreenStepType } from "./NewCustomScreen/types";
+import { ComposableScreenStepType } from "./ComposableScreen/types";
 import { PickerStepType } from "./Picker/types";
 import { QuestionStepType } from "./Question/types";
 import { RatingsStepType } from "./Ratings/types";
@@ -12,7 +12,7 @@ export type OnboardingStepType =
   | CommitmentStepType
   | LoaderStepType
   | MediaContentStepType
-  | NewCustomScreenStepType
+  | ComposableScreenStepType
   | PickerStepType
   | QuestionStepType
   | RatingsStepType;
